refactor: remove useless parser and use `?` where possible
